Address C comments and React.use..
