+++ docs/src/jsx/commands/DuckScene.js
@@ -7,7 +7,7 @@
 // #BEGIN EXAMPLE
 import React from "react";
 
-import duckModel from "./utils/Duck.glb"; // URL pointing to a .glb file
+import duckModel from "../utils/Duck.glb"; // URL pointing to a .glb file
 import Worldview, { Axes, Grid, GLTFScene, DEFAULT_CAMERA_STATE } from "regl-worldview";
 
 // #BEGIN EDITABLE
