+++ docs/src/jsx/tutorials/ColorfulKnot.js
