+++ docs/src/jsx/tutorials/FollowObject.js
@@ -5,26 +5,23 @@
 //  You may not use this file except in compliance with the License.
 
 // #BEGIN EXAMPLE
-import React from "react";
+import React, { useState } from "react";
 
 import duckModel from "../utils/Duck.glb";
 import useRequestAnimationFrame from "../utils/useRequestAnimationFrame";
-import Worldview, { Cubes, Spheres, Axes, GLTFScene, DEFAULT_CAMERA_STATE } from "regl-worldview";
+import Worldview, { Spheres, Axes, GLTFScene, DEFAULT_CAMERA_STATE } from "regl-worldview";
 
 // #BEGIN EDITABLE
 function Example() {
   const steps = 500; // total amount of objects
-  const [clickedObjectIds, setClickedObjectIds] = React.useState([]);
-  const [count, setCount] = React.useState(0);
-  const [shouldStopDuck, setShouldStopDuck] = React.useState(false);
-
+  const [count, setCount] = useState(0);
   useRequestAnimationFrame(
     () => {
       // update count before each browser repaint
       const newCount = (count + 1) % steps;
       setCount(newCount);
     },
-    shouldStopDuck, // disable the animation when the duck is not moving
+    false,
     []
   );
 
@@ -40,12 +37,11 @@ function Example() {
   // the object index needs to multiple by this scale so it's evenly distributed in the space
   const scale = (Math.PI * 2) / steps;
   const sphereMarker = {
-    id: 1001,
     pose: {
       orientation: { x: 0, y: 0, z: 0, w: 1 },
       position: { x: 0, y: 0, z: 0 },
     },
-    scale: { x: 3, y: 3, z: 3 },
+    scale: { x: 1, y: 1, z: 1 },
     colors: [],
     points: [],
   };
@@ -63,59 +59,21 @@ function Example() {
       sphereMarker.colors.push(numberToColor(idx, steps));
       sphereMarker.points.push({ x: x * 20, y: y * 20, z: z * 20 });
     });
-
-  const obstacleMarkers = Array.from(clickedObjectIds).map((clickedObjectId, index) => {
-    const pointIdx = clickedObjectId - sphereMarker.id;
-    const position = sphereMarker.points[pointIdx];
-    return {
-      // Since the `sphereMarker` has used up the id range: 101 ~ 101 + 499 (inclusive, each id represent one sphere object),
-      // to make the obstacleMarkers' ids unique, we'll use the range: 500 (sphereMarker.id + step) + index.
-      // Learn about id mapping at https://cruise-automation.github.io/webviz/worldview/#/docs/api/mouse-events
-      id: sphereMarker.id + steps + index,
-      // remember the original clickedObjectId so when the obstacle is clicked, we can
-      // remove the obstacle quickly by updating clickedObjectIds
-      clickedObjectId,
-      pose: {
-        orientation: { x: 0, y: 0, z: 0, w: 1 },
-        position,
-      },
-      color: { r: 1, g: 0, b: 0, a: 1 }, // red
-      scale: { x: 6, y: 6, z: 6 }, // scale up a little so it's bigger than the spheres
-    };
-  });
   const duckPosition = sphereMarker.points[count];
-  // make the duck stop
-  React.useEffect(
-    () => {
-      const duckPositionId = sphereMarker.id + count;
-      if (!shouldStopDuck && clickedObjectIds.includes(duckPositionId)) {
-        setShouldStopDuck(true);
-      }
-    },
-    [clickedObjectIds, shouldStopDuck, count, sphereMarker.id]
-  );
 
   return (
     <Worldview
-      defaultCameraState={{
+      cameraState={{
+        // Default setting for cameraState.
+        // Learn more at https://cruise-automation.github.io/webviz/worldview/#/docs/api/camera
         ...DEFAULT_CAMERA_STATE,
-        distance: 160,
+        // This is the magic! Simply supply the target position and the camera will follow
+        target: [duckPosition.x, duckPosition.y, duckPosition.z],
         thetaOffset: -Math.PI / 2, // rotate the camera so the duck is facing right
+        // zoom out a little so we can see better
+        distance: 160,
       }}>
-      <Spheres
-        onClick={(ev, { objectId }) => {
-          setClickedObjectIds([...clickedObjectIds, objectId]);
-        }}>
-        {[sphereMarker]}
-      </Spheres>
-      <Cubes
-        onClick={(ev, { object }) => {
-          const newClickedObjectIds = clickedObjectIds.filter((id) => id !== object.clickedObjectId);
-          setClickedObjectIds(newClickedObjectIds);
-          setShouldStopDuck(false);
-        }}>
-        {obstacleMarkers}
-      </Cubes>
+      <Spheres>{[sphereMarker]}</Spheres>
       <Axes />
       {/* Download model: https://github.com/cruise-automation/webviz/blob/master/docs/src/jsx/utils/Duck.glb  */}
       <GLTFScene model={duckModel}>
@@ -124,7 +82,7 @@ function Example() {
             position: duckPosition,
             orientation: { x: 0, y: 0, z: 0, w: 1 },
           },
-          scale: { x: 4, y: 4, z: 4 },
+          scale: { x: 3, y: 3, z: 3 },
         }}
       </GLTFScene>
     </Worldview>
