+++ docs/src/jsx/tutorials/FollowObjectOrientation.js
@@ -5,7 +5,8 @@
 //  You may not use this file except in compliance with the License.
 
 // #BEGIN EXAMPLE
-import React from "react";
+import { quat, vec3 } from "gl-matrix";
+import React, { useState } from "react";
 
 import duckModel from "../utils/Duck.glb";
 import useRequestAnimationFrame from "../utils/useRequestAnimationFrame";
@@ -14,7 +15,7 @@ import Worldview, { Spheres, Axes, GLTFScene, DEFAULT_CAMERA_STATE } from "regl-
 // #BEGIN EDITABLE
 function Example() {
   const steps = 500; // total amount of objects
-  const [count, setCount] = React.useState(0);
+  const [count, setCount] = useState(0);
   useRequestAnimationFrame(
     () => {
       // update count before each browser repaint
@@ -61,17 +62,35 @@ function Example() {
     });
   const duckPosition = sphereMarker.points[count];
 
+  // get the orientation for the duck so its moving direction is always aligned with the knot
+  const tangentVecO = [
+    Math.cos(count * scale) + 4 * Math.cos(2 * count * scale),
+    -Math.sin(count * scale) + 4 * Math.sin(2 * count * scale),
+    -3 * Math.cos(3 * count * scale),
+  ];
+  const tangentVec = vec3.normalize([0, 0, 0], tangentVecO);
+  const duckYaw = quat.setAxisAngle([0, 0, 0, 0], [0, 0, 1], Math.atan2(tangentVec[1], tangentVec[0]));
+  const duckOrientation = [0, 0, 0, 1];
+  quat.multiply(duckOrientation, duckOrientation, duckYaw);
+  quat.rotateY(
+    duckOrientation,
+    duckOrientation,
+    -Math.atan(tangentVecO[2] / Math.hypot(tangentVecO[0], tangentVecO[1]))
+  );
+  quat.rotateZ(duckOrientation, duckOrientation, -Math.PI / 2);
+
   return (
     <Worldview
       cameraState={{
         // Default setting for cameraState.
         // Learn more at https://cruise-automation.github.io/webviz/worldview/#/docs/api/camera
         ...DEFAULT_CAMERA_STATE,
-        // This is the magic! Simply supply the target position and the camera will follow
         target: [duckPosition.x, duckPosition.y, duckPosition.z],
-        thetaOffset: -Math.PI / 2, // rotate the camera so the duck is facing right
+        // This is the magic! The `targetOrientation` input will make sure the camera follows the duck's orientation
+        targetOrientation: duckOrientation,
         // zoom out a little so we can see better
         distance: 160,
+        thetaOffset: -Math.PI / 2, // rotate the camera so the duck is facing right
       }}>
       <Spheres>{[sphereMarker]}</Spheres>
       <Axes />
@@ -80,7 +99,7 @@ function Example() {
         {{
           pose: {
             position: duckPosition,
-            orientation: { x: 0, y: 0, z: 0, w: 1 },
+            orientation: { x: duckOrientation[0], y: duckOrientation[1], z: duckOrientation[2], w: duckOrientation[3] },
           },
           scale: { x: 3, y: 3, z: 3 },
         }}
