+++ docs/src/jsx/tutorials/InstancedRendering.js
