+++ docs/src/jsx/utils/useRequestAnimationFrame.js
@@ -4,15 +4,15 @@
 //  found in the LICENSE file in the root directory of this source tree.
 //  You may not use this file except in compliance with the License.
 
-import React from "react";
+import { useRef, useEffect } from "react";
 
 // a react hook that accepts a callback function which will be called before each browser repaint
 export default function useRequestAnimationFrame(callback, disable, dependencies) {
   // persist the requestAnimationFrameId so the last refresh callback can be cancelled when component unmounts
-  const rafId = React.useRef();
+  const rafId = useRef();
   // start the loop when the component mounts and cancel the animation frame when unmounts
   // or related dependency changes
-  React.useEffect(
+  useEffect(
     () => {
       if (!disable) {
         // eslint-disable-next-line no-inner-declarations
